+++ R/bin_columns_into_quantiles.r
@@ -1,37 +1,36 @@
 #' \code{bin_columns_into_quantiles}
 #'
 #' \code{bin_columns_into_quantiles} is an internal function used to convert a
-#' matrix of specificity (with columns of cell types) intom a matrix of
-#' specificity quantiles
-#'
-#' @param matrixIn The matrix of specificity values
-#' @param numberOfBins Number of quantile 'bins' to use (40 is recommended)
+#' vector of specificity  into a vector of specificity quantiles.
+#' This function can be iterated across a matrix using \link[base]{apply} 
+#' to create a matrix of specificity quantiles.
+#' @param vec The vector of gene of specificity values.
+#' @param numberOfBins Number of quantile bins to use (40 is recommended).
 #' @param defaultBin Which bin to assign when there's only one
 #' non-zero quantile. In situations where there's only one non-zero quantile,
-#' \code{cut()} throws an error. Avoid these situations by
+#' \link[base]{cut} throws an error. Avoid these situations by
 #' using a default quantile.
-#' @return A matrix with same shape as matrixIn but with columns storing
-#' quantiles instead of specificity
+#' @returns A vector with same length as \code{vec} but with columns storing
+#' quantiles instead of specificity.
 #' @examples
 #' ctd <- ewceData::ctd()
 #' ctd[[1]]$specificity_quantiles <- apply(ctd[[1]]$specificity, 2,
-#'     FUN = bin_columns_into_quantiles,
-#'     numberOfBins = 40
-#' )
+#'     FUN = bin_columns_into_quantiles)
 #' @export
 #' @importFrom stats quantile
-bin_columns_into_quantiles <- function(matrixIn,
+bin_columns_into_quantiles <- function(vec,
                                        numberOfBins = 40,
                                        defaultBin = as.integer(
                                            numberOfBins / 2)
                                        ) {
-    quantileValues <- rep(0, length(matrixIn))
-    breaks <- unique(stats::quantile(matrixIn[matrixIn > 0],
+    
+    quantileValues <- rep(0, length(vec))
+    breaks <- unique(stats::quantile(vec[vec > 0],
         probs = seq(0, 1, by = 1 / numberOfBins),
         na.rm = TRUE
-    ))
+    )) 
     if (length(breaks) > 1) {
-        quantileValues[matrixIn > 0] <- as.numeric(cut(matrixIn[matrixIn > 0],
+        quantileValues[vec > 0] <- as.numeric(cut(vec[vec > 0],
             breaks = breaks,
             include.lowest = TRUE
         ))
@@ -44,7 +43,7 @@ bin_columns_into_quantiles <- function(matrixIn,
             "Assigning these values to default quantile ",
             "(", defaultBin, ")"
         )
-        quantileValues[matrixIn > 0] <- defaultBin
+        quantileValues[vec > 0] <- defaultBin
     }
     return(quantileValues)
 }